feat: Added Azure Naming Service
