+++ src/services/resourceService.ts
@@ -2,6 +2,7 @@ import Serverless from "serverless";
 import { ResourceManagementClient } from "@azure/arm-resources";
 import { BaseService } from "./baseService";
 import { Utils } from "../shared/utils";
+import { AzureNamingService } from "./namingService";
 
 export class ResourceService extends BaseService {
   private resourceClient: ResourceManagementClient;
@@ -57,7 +58,7 @@ export class ResourceService extends BaseService {
     this.log(`Creating resource group: ${this.resourceGroup}`);
 
     return await this.resourceClient.resourceGroups.createOrUpdate(this.resourceGroup, {
-      location: Utils.getNormalizedRegionName(this.getRegion()),
+      location: AzureNamingService.getNormalizedRegionName(this.getRegion()),
     });
   }
