+++ src/shared/utils.ts
@@ -1,9 +1,8 @@
 import { relative } from "path";
 import Serverless from "serverless";
+import { ServerlessAzureFunctionConfig } from "../models/serverless";
 import { BindingUtils } from "./bindings";
 import { constants } from "./constants";
-import { Guard } from "./guard";
-import { ServerlessAzureFunctionConfig } from "../models/serverless";
 
 export interface FunctionMetadata {
   entryPoint: any;
@@ -130,72 +129,6 @@ export class Utils {
       result += (s.substr(0, substringSize));
     }
     return result;
-  }
-
-  /**
-   * Creates a short name to be used for state name abbreviation
-   * @param stageName The stage name
-   */
-  public static createShortStageName(stageName: string) {
-    Guard.empty(stageName);
-
-    const stageMap = {
-      "dogfood": "df",
-      "production": "prod",
-      "development": "dev",
-      "testing": "test"
-    };
-
-    return this.createShortName(stageName, stageMap);
-  }
-
-  /**
-   * Gets the normalized region name from long name (ex. West US 2 -> westus2)
-   * @param regionName The region name
-   */
-  public static getNormalizedRegionName(regionName: string) {
-    Guard.empty(regionName);
-    return regionName.replace(/\W/g, "").toLowerCase();
-  }
-
-  /**
-   * Creates a short name for an azure region
-   * @param regionName The azure region name
-   */
-  public static createShortAzureRegionName(regionName: string) {
-    Guard.empty(regionName);
-
-    const locationMap = {
-      "north": "n",
-      "south": "s",
-      "east": "e",
-      "west": "w",
-      "central": "c",
-    };
-
-    return this.createShortName(regionName, locationMap);
-  }
-
-  /**
-   * Creates a short name from a long name based on a well-known string map
-   * @param longName The long name to replace
-   * @param wellKnownMap A well known map of long terms to short abbreviations
-   */
-  private static createShortName(longName: string, wellKnownMap: { [key: string]: string }) {
-    Guard.empty(longName);
-    Guard.null(wellKnownMap);
-
-    const pattern = `(${Object.keys(wellKnownMap).join("|")})`;
-    const regex = new RegExp(pattern, "g");
-
-    return longName
-      .replace(/\W+/g, "")
-      .toLowerCase()
-      .split(regex)
-      .map((part) => {
-        return wellKnownMap[part] || part.substr(0, 3);
-      })
-      .join("");
   }
 
   public static get(object: any, key: string, defaultValue?: any) {
